some improvements in day01::ex05 {dispatching} and day01::ex06 {SWITCH statement}

+++ Day05/ex03/Intern.cpp
@@ -64,7 +64,7 @@ AForm*	Intern::makeForm(std::string const formName, std::string const formTarget
 		"presidential pardon"
 	};
 
-	for (int index = 0; index < 3; index++)
+	for (size_t index = 0; index < 3; index++)
 	{
 		if (formName == arrayOfString[index])
 			return ((this->*arrayOfMemberFunPtr[index])(formTarget));
